Fixed setup file error
It was noticed that the setup file kept erroring out after it would
close. The twitter app file WOULD be created, however, the script would
not be able to continue. I suspected that this was either a logical or
buffer issue.

I’ve added python’s .flush() as well as used fsync() to try and combat
potential buffer issues and am now using Try, Except, Finally to
possibly squash logic issues.