Save file with name of filter
Saving the file as STREAM.csv from the twitter portion is great and
all, but, what if you have a bunch of streams that are all separate?
Well, now it appends the name of the filter used to the csv name.

So, if you filter by “hacking” is will save output of the twitter
streaming output as STREAM_hacking.csv.

Love,
NinjaSl0th